feat(MovieRepository): 영화를 예매할 때 어떤 극장에서 예매했는지도 알려준다

+++ domain/src/main/java/woowacourse/movie/model/Price.kt
@@ -1,7 +1,7 @@
 package woowacourse.movie.model
 
 @JvmInline
-value class Price(val price: Long) {
+value class Price(val price: Int) {
     init {
         require(price >= MIN_PRICE) {
             "$price - 가격은 ${MIN_PRICE}원 이상 이어야 합니다."
